Polished and removed excess code

+++ src/model/LamportClock.java
@@ -1,13 +1,13 @@
 package model;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class LamportClock {
-    private final AtomicInteger time;
+    private final AtomicLong time;
     private final int nodeIncrement;
 
     public LamportClock(int nodeIncrement) {
-        this.time = new AtomicInteger(0);
+        this.time = new AtomicLong(0);
         this.nodeIncrement = nodeIncrement;
     }
 
@@ -15,14 +15,14 @@ public class LamportClock {
         time.getAndAdd(nodeIncrement);
     }
 
-    public void update(int receivedTime) {
-        int currentTime;
+    public void update(long receivedTime) {
+        long currentTime;
         do {
             currentTime = time.get();
         } while (!time.compareAndSet(currentTime, Math.max(currentTime, receivedTime) + nodeIncrement));
     }
 
-    public int getTime() {
+    public long getTime() {
         return time.get();
     }
 }
